refactor: 딜러와 플레이어 사이의 중복 코드를 추상화를 사용해 제거
