+++ src/main/java/blackjack/domain/Player.java
@@ -1,20 +1,17 @@
 package blackjack.domain;
 
 import java.util.Collections;
-import java.util.List;
 import java.util.Set;
 
-public class Player {
+public class Player extends Participant {
     private final String name;
-    private final CardDeck cardDeck;
-    private final CardDump cardDump;
 
     public Player(String name, CardDeck cardDeck, CardDump cardDump) {
+        super(cardDeck, cardDump);
         this.name = name;
-        this.cardDeck = cardDeck;
-        this.cardDump = cardDump;
     }
 
+    @Override
     public boolean canTakeExtraCard() {
         Set<Integer> possibleSum = cardDeck.calculatePossibleSum();
         int minScore = Collections.min(possibleSum);
@@ -22,6 +19,7 @@ public class Player {
         return minScore <= 21;
     }
 
+    @Override
     public int calculateTotalCardScore() {
         Set<Integer> possibleSum = cardDeck.calculatePossibleSum();
         return possibleSum.stream()
@@ -30,22 +28,7 @@ public class Player {
                 .orElse(Collections.min(possibleSum));
     }
 
-    public boolean isBust() {
-        Set<Integer> possibleSum = cardDeck.calculatePossibleSum();
-        int totalScore = Collections.max(possibleSum);
-        return totalScore > 21;
-    }
-
-    public void addCard() {
-        Card card = cardDump.drawCard();
-        cardDeck.add(card);
-    }
-
     public String getName() {
         return name;
-    }
-
-    public List<Card> getCardDeck() {
-        return cardDeck.getCards();
     }
 }
